refactor: move components, since they're gonna be used in other parts of app
